refactor: split linker admin tasks
