+++ handlers/linker/bulk_approve_task.go
@@ -1,0 +1,103 @@
+package linker
+
+import (
+	"database/sql"
+	"fmt"
+	"log"
+	"net/http"
+	"strconv"
+	"strings"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/internal/db"
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+	"github.com/arran4/goa4web/workers/searchworker"
+)
+
+// bulkApproveTask approves multiple queued linker items.
+type bulkApproveTask struct{ tasks.TaskString }
+
+var BulkApproveTask = &bulkApproveTask{TaskString: TaskBulkApprove}
+
+var (
+	_ tasks.Task                                    = (*bulkApproveTask)(nil)
+	_ notif.SubscribersNotificationTemplateProvider = (*bulkApproveTask)(nil)
+	_ notif.AdminEmailTemplateProvider              = (*bulkApproveTask)(nil)
+	_ searchworker.IndexedTask                      = bulkApproveTask{}
+)
+
+func (bulkApproveTask) IndexType() string { return searchworker.TypeLinker }
+
+func (bulkApproveTask) IndexData(data map[string]any) []searchworker.IndexEventData {
+	if v, ok := data[searchworker.EventKey].(searchworker.IndexEventData); ok {
+		return []searchworker.IndexEventData{v}
+	}
+	return nil
+}
+
+func (bulkApproveTask) Action(w http.ResponseWriter, r *http.Request) any {
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+	if err := r.ParseForm(); err != nil {
+		log.Printf("ParseForm Error: %s", err)
+	}
+	var links []map[string]any
+	for _, q := range r.Form["qid"] {
+		id, _ := strconv.Atoi(q)
+		lid, err := queries.SelectInsertLInkerQueuedItemIntoLinkerByLinkerQueueId(r.Context(), int32(id))
+		if err != nil {
+			log.Printf("selectInsert Error: %s", err)
+			continue
+		}
+		cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
+		link, err := queries.GetLinkerItemByIdWithPosterUsernameAndCategoryTitleDescendingForUser(r.Context(), db.GetLinkerItemByIdWithPosterUsernameAndCategoryTitleDescendingForUserParams{
+			ViewerID:     cd.UserID,
+			Idlinker:     int32(lid),
+			ViewerUserID: sql.NullInt32{Int32: cd.UserID, Valid: cd.UserID != 0},
+		})
+		if err != nil {
+			log.Printf("getLinkerItemById Error: %s", err)
+			continue
+		}
+		text := strings.Join([]string{link.Title.String, link.Description.String}, " ")
+		if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
+			if evt := cd.Event(); evt != nil {
+				if evt.Data == nil {
+					evt.Data = map[string]any{}
+				}
+				u, _ := cd.CurrentUser()
+				mod := ""
+				if u != nil {
+					mod = u.Username.String
+				}
+				links = append(links, map[string]any{"Title": link.Title.String, "URL": link.Url.String, "Username": link.Username.String, "Moderator": mod})
+				evt.Data["Moderator"] = mod
+				evt.Data["links"] = links
+				if len(links) == 1 {
+					evt.Data["LinkURL"] = cd.AbsoluteURL(fmt.Sprintf("/linker/show/%d", lid))
+				}
+				evt.Data[searchworker.EventKey] = searchworker.IndexEventData{Type: searchworker.TypeLinker, ID: int32(lid), Text: text}
+			}
+		}
+	}
+	return nil
+}
+
+func (bulkApproveTask) SubscribedEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("linkerApprovedEmail")
+}
+
+func (bulkApproveTask) SubscribedInternalNotificationTemplate() *string {
+	s := notif.NotificationTemplateFilenameGenerator("linker_approved")
+	return &s
+}
+
+func (bulkApproveTask) AdminEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("adminNotificationLinkerApprovedEmail")
+}
+
+func (bulkApproveTask) AdminInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("adminNotificationLinkerApprovedEmail")
+	return &v
+}
