+++ handlers/linker/linkerAdminQueuePage.go
@@ -14,10 +14,6 @@ import (
 	"github.com/arran4/goa4web/core/common"
 	"github.com/arran4/goa4web/handlers"
 	"github.com/arran4/goa4web/internal/db"
-	notif "github.com/arran4/goa4web/internal/notifications"
-	"github.com/arran4/goa4web/workers/searchworker"
-
-	"github.com/arran4/goa4web/internal/tasks"
 )
 
 func AdminQueuePage(w http.ResponseWriter, r *http.Request) {
@@ -119,71 +115,6 @@ func AdminQueuePage(w http.ResponseWriter, r *http.Request) {
 	handlers.TemplateHandler(w, r, "adminQueuePage.gohtml", data)
 }
 
-type deleteTask struct{ tasks.TaskString }
-
-var DeleteTask = &deleteTask{TaskString: TaskDelete}
-var _ tasks.Task = (*deleteTask)(nil)
-
-var (
-	_ tasks.Task                                    = (*deleteTask)(nil)
-	_ notif.SubscribersNotificationTemplateProvider = (*deleteTask)(nil)
-	_ notif.AdminEmailTemplateProvider              = (*deleteTask)(nil)
-)
-
-func (deleteTask) Action(w http.ResponseWriter, r *http.Request) any {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	qid, _ := strconv.Atoi(r.URL.Query().Get("qid"))
-	var link *db.GetAllLinkerQueuedItemsWithUserAndLinkerCategoryDetailsRow
-	if rows, err := queries.GetAllLinkerQueuedItemsWithUserAndLinkerCategoryDetails(r.Context()); err == nil {
-		for _, it := range rows {
-			if it.Idlinkerqueue == int32(qid) {
-				link = it
-				break
-			}
-		}
-	}
-	if err := queries.DeleteLinkerQueuedItem(r.Context(), int32(qid)); err != nil {
-		return fmt.Errorf("delete linker queued item fail %w", handlers.ErrRedirectOnSamePageHandler(err))
-	}
-	if link != nil {
-		if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-			if evt := cd.Event(); evt != nil {
-				if evt.Data == nil {
-					evt.Data = map[string]any{}
-				}
-				u, _ := cd.CurrentUser()
-				mod := ""
-				if u != nil {
-					mod = u.Username.String
-				}
-				evt.Data["Title"] = link.Title.String
-				evt.Data["Username"] = link.Username.String
-				evt.Data["Moderator"] = mod
-				evt.Data["LinkURL"] = link.Url.String
-			}
-		}
-	}
-	return nil
-}
-
-func (deleteTask) SubscribedEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("linkerRejectedEmail")
-}
-
-func (deleteTask) SubscribedInternalNotificationTemplate() *string {
-	s := notif.NotificationTemplateFilenameGenerator("linker_rejected")
-	return &s
-}
-
-func (deleteTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationLinkerRejectedEmail")
-}
-
-func (deleteTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationLinkerRejectedEmail")
-	return &v
-}
-
 func AdminQueueUpdateActionPage(w http.ResponseWriter, r *http.Request) {
 	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
 	qid, _ := strconv.Atoi(r.URL.Query().Get("qid"))
@@ -203,246 +134,4 @@ func AdminQueueUpdateActionPage(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	handlers.TaskDoneAutoRefreshPage(w, r)
-}
-
-type approveTask struct{ tasks.TaskString }
-
-var ApproveTask = &approveTask{TaskString: TaskApprove}
-var _ tasks.Task = (*approveTask)(nil)
-
-var (
-	_ tasks.Task                                    = (*approveTask)(nil)
-	_ notif.SubscribersNotificationTemplateProvider = (*approveTask)(nil)
-	_ notif.AdminEmailTemplateProvider              = (*approveTask)(nil)
-	_ searchworker.IndexedTask                      = approveTask{}
-)
-
-func (approveTask) IndexType() string { return searchworker.TypeLinker }
-
-func (approveTask) IndexData(data map[string]any) []searchworker.IndexEventData {
-	if v, ok := data[searchworker.EventKey].(searchworker.IndexEventData); ok {
-		return []searchworker.IndexEventData{v}
-	}
-	return nil
-}
-
-func (approveTask) Action(w http.ResponseWriter, r *http.Request) any {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	qid, _ := strconv.Atoi(r.URL.Query().Get("qid"))
-	lid, err := queries.SelectInsertLInkerQueuedItemIntoLinkerByLinkerQueueId(r.Context(), int32(qid))
-	if err != nil {
-		return fmt.Errorf("approve linker item fail %w", handlers.ErrRedirectOnSamePageHandler(err))
-	}
-
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-	link, err := queries.GetLinkerItemByIdWithPosterUsernameAndCategoryTitleDescendingForUser(r.Context(), db.GetLinkerItemByIdWithPosterUsernameAndCategoryTitleDescendingForUserParams{
-		ViewerID:     cd.UserID,
-		Idlinker:     int32(lid),
-		ViewerUserID: sql.NullInt32{Int32: cd.UserID, Valid: cd.UserID != 0},
-	})
-	if err != nil {
-		return fmt.Errorf("get linker item fail %w", handlers.ErrRedirectOnSamePageHandler(err))
-	}
-
-	text := strings.Join([]string{link.Title.String, link.Description.String}, " ")
-	if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-		if evt := cd.Event(); evt != nil {
-			if evt.Data == nil {
-				evt.Data = map[string]any{}
-			}
-			u, _ := cd.CurrentUser()
-			mod := ""
-			if u != nil {
-				mod = u.Username.String
-			}
-			evt.Data["Title"] = link.Title.String
-			evt.Data["Username"] = link.Username.String
-			evt.Data["Moderator"] = mod
-			evt.Data["LinkURL"] = cd.AbsoluteURL(fmt.Sprintf("/linker/show/%d", lid))
-			evt.Data[searchworker.EventKey] = searchworker.IndexEventData{Type: searchworker.TypeLinker, ID: int32(lid), Text: text}
-		}
-	}
-	return nil
-}
-
-type bulkDeleteTask struct{ tasks.TaskString }
-
-var BulkDeleteTask = &bulkDeleteTask{TaskString: TaskBulkDelete}
-
-var (
-	_ tasks.Task                                    = (*bulkDeleteTask)(nil)
-	_ notif.SubscribersNotificationTemplateProvider = (*bulkDeleteTask)(nil)
-	_ notif.AdminEmailTemplateProvider              = (*bulkDeleteTask)(nil)
-)
-
-func (bulkDeleteTask) Action(w http.ResponseWriter, r *http.Request) any {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	if err := r.ParseForm(); err != nil {
-		log.Printf("ParseForm Error: %s", err)
-	}
-	var info []map[string]any
-	if rows, err := queries.GetAllLinkerQueuedItemsWithUserAndLinkerCategoryDetails(r.Context()); err == nil {
-		ids := make(map[int]struct{})
-		for _, q := range r.Form["qid"] {
-			id, _ := strconv.Atoi(q)
-			ids[id] = struct{}{}
-		}
-		for _, it := range rows {
-			if _, ok := ids[int(it.Idlinkerqueue)]; ok {
-				info = append(info, map[string]any{"Title": it.Title.String, "URL": it.Url.String, "Username": it.Username.String})
-			}
-		}
-	}
-	for _, q := range r.Form["qid"] {
-		id, _ := strconv.Atoi(q)
-		if err := queries.DeleteLinkerQueuedItem(r.Context(), int32(id)); err != nil {
-			log.Printf("deleteLinkerQueuedItem Error: %s", err)
-		}
-	}
-	if len(info) > 0 {
-		if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-			if evt := cd.Event(); evt != nil {
-				if evt.Data == nil {
-					evt.Data = map[string]any{}
-				}
-				u, _ := cd.CurrentUser()
-				mod := ""
-				if u != nil {
-					mod = u.Username.String
-				}
-				for i := range info {
-					info[i]["Moderator"] = mod
-				}
-				evt.Data["links"] = info
-				if len(info) == 1 {
-					if url, ok := info[0]["URL"].(string); ok {
-						evt.Data["LinkURL"] = url
-					}
-				}
-				evt.Data["Moderator"] = mod
-			}
-		}
-	}
-	return nil
-}
-
-func (bulkDeleteTask) SubscribedEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("linkerRejectedEmail")
-}
-
-func (bulkDeleteTask) SubscribedInternalNotificationTemplate() *string {
-	s := notif.NotificationTemplateFilenameGenerator("linker_rejected")
-	return &s
-}
-
-func (bulkDeleteTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationLinkerRejectedEmail")
-}
-
-func (bulkDeleteTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationLinkerRejectedEmail")
-	return &v
-}
-
-func (approveTask) SubscribedEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("linkerApprovedEmail")
-}
-
-func (approveTask) SubscribedInternalNotificationTemplate() *string {
-	s := notif.NotificationTemplateFilenameGenerator("linker_approved")
-	return &s
-}
-
-func (approveTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationLinkerApprovedEmail")
-}
-
-func (approveTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationLinkerApprovedEmail")
-	return &v
-}
-
-type bulkApproveTask struct{ tasks.TaskString }
-
-var BulkApproveTask = &bulkApproveTask{TaskString: TaskBulkApprove}
-
-var (
-	_ tasks.Task                                    = (*bulkApproveTask)(nil)
-	_ notif.SubscribersNotificationTemplateProvider = (*bulkApproveTask)(nil)
-	_ notif.AdminEmailTemplateProvider              = (*bulkApproveTask)(nil)
-	_ searchworker.IndexedTask                      = bulkApproveTask{}
-)
-
-func (bulkApproveTask) IndexType() string { return searchworker.TypeLinker }
-
-func (bulkApproveTask) IndexData(data map[string]any) []searchworker.IndexEventData {
-	if v, ok := data[searchworker.EventKey].(searchworker.IndexEventData); ok {
-		return []searchworker.IndexEventData{v}
-	}
-	return nil
-}
-
-func (bulkApproveTask) Action(w http.ResponseWriter, r *http.Request) any {
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	if err := r.ParseForm(); err != nil {
-		log.Printf("ParseForm Error: %s", err)
-	}
-	var links []map[string]any
-	for _, q := range r.Form["qid"] {
-		id, _ := strconv.Atoi(q)
-		lid, err := queries.SelectInsertLInkerQueuedItemIntoLinkerByLinkerQueueId(r.Context(), int32(id))
-		if err != nil {
-			log.Printf("selectInsert Error: %s", err)
-			continue
-		}
-		cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-		link, err := queries.GetLinkerItemByIdWithPosterUsernameAndCategoryTitleDescendingForUser(r.Context(), db.GetLinkerItemByIdWithPosterUsernameAndCategoryTitleDescendingForUserParams{
-			ViewerID:     cd.UserID,
-			Idlinker:     int32(lid),
-			ViewerUserID: sql.NullInt32{Int32: cd.UserID, Valid: cd.UserID != 0},
-		})
-		if err != nil {
-			log.Printf("getLinkerItemById Error: %s", err)
-			continue
-		}
-		text := strings.Join([]string{link.Title.String, link.Description.String}, " ")
-		if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
-			if evt := cd.Event(); evt != nil {
-				if evt.Data == nil {
-					evt.Data = map[string]any{}
-				}
-				u, _ := cd.CurrentUser()
-				mod := ""
-				if u != nil {
-					mod = u.Username.String
-				}
-				links = append(links, map[string]any{"Title": link.Title.String, "URL": link.Url.String, "Username": link.Username.String, "Moderator": mod})
-				evt.Data["Moderator"] = mod
-				evt.Data["links"] = links
-				if len(links) == 1 {
-					evt.Data["LinkURL"] = cd.AbsoluteURL(fmt.Sprintf("/linker/show/%d", lid))
-				}
-				evt.Data[searchworker.EventKey] = searchworker.IndexEventData{Type: searchworker.TypeLinker, ID: int32(lid), Text: text}
-			}
-		}
-	}
-	return nil
-}
-
-func (bulkApproveTask) SubscribedEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("linkerApprovedEmail")
-}
-
-func (bulkApproveTask) SubscribedInternalNotificationTemplate() *string {
-	s := notif.NotificationTemplateFilenameGenerator("linker_approved")
-	return &s
-}
-
-func (bulkApproveTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationLinkerApprovedEmail")
-}
-
-func (bulkApproveTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationLinkerApprovedEmail")
-	return &v
 }
